requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
lockfile==0.12.2
python-daemon==3.1.2
requests==2.32.3
urllib3==2.4.0
wcwidth==0.2.13
